Move classes into facilitymanagement package
Refactor project structure by moving resource-related classes and configs into the Backend.*.facilitymanagement packages. Updated package declarations and imports for SecurityConfig, WebConfig, ResourceController, Resource, ResourceRepository, and ResourceService to reflect the new paths.
