+++ Backend/src/main/java/Backend/repository/facilitymanagement/ResourceRepository.java
@@ -1,6 +1,6 @@
-package Backend.repository;
+package Backend.repository.facilitymanagement;
 
-import Backend.model.Resource;
+import Backend.model.facilitymanagement.Resource;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
